requirements.txt
attrs==23.2.0
aws-cdk-lib==2.138.0
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.3
aws-cdk.integ-tests-alpha==2.135.0a0
aws-solutions-constructs.aws-lambda-stepfunctions==2.55.0
aws-solutions-constructs.aws-s3-lambda==2.55.0
aws-solutions-constructs.core==2.55.0
boto3==1.34.87
botocore==1.34.87
cattrs==23.2.3
cdk-nag==2.28.93
constructs==10.3.0
exceptiongroup==1.2.1
importlib_resources==6.4.0
jmespath==1.0.1
jsii==1.97.0
publication==0.0.3
python-dateutil==2.9.0.post0
pytz==2024.1
s3transfer==0.10.1
six==1.16.0
typeguard==2.13.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
zipp==3.18.1
